Default transfer is now in bytes now gigabytes

+++ src/main/java/com/hubert/downloader/domain/InformationSize.java
@@ -4,20 +4,20 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 
-public record InformationSize(InformationUnit unit, Float size) {
+public record InformationSize(InformationUnit unit, Long size) {
 
     public InformationSize parseTo(InformationUnit toUnit) {
         List<InformationUnit> units = Arrays.asList(InformationUnit.values());
 
-        Float sizeOf = units.indexOf(unit) > units.indexOf(toUnit)
+        Long sizeOf = units.indexOf(unit) > units.indexOf(toUnit)
                 ? parseSizeInDescending(units, toUnit)
                 : parseInNormalOrder(units, toUnit);
 
         return new InformationSize(toUnit, sizeOf);
     }
 
-    private Float parseSizeInDescending(List<InformationUnit> units, InformationUnit toUnit) {
-        Float sizeOf = size;
+    private Long parseSizeInDescending(List<InformationUnit> units, InformationUnit toUnit) {
+        Long sizeOf = size;
 
         Collections.reverse(units);
 
@@ -30,8 +30,8 @@ public record InformationSize(InformationUnit unit, Float size) {
         return sizeOf;
     }
 
-    private Float parseInNormalOrder(List<InformationUnit> units, InformationUnit toUnit) {
-        Float sizeOf = size;
+    private Long parseInNormalOrder(List<InformationUnit> units, InformationUnit toUnit) {
+        Long sizeOf = size;
 
         for (InformationUnit unit : units.subList(units.indexOf(unit), units.size()-1)) {
             if (unit.equals(toUnit)) break;
