+++ src/main/java/com/hubert/downloader/domain/Transfer.java
@@ -1,6 +1,7 @@
 package com.hubert.downloader.domain;
 
 import lombok.*;
+import lombok.extern.java.Log;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 @Getter
@@ -14,10 +15,10 @@ public class Transfer {
     private InformationSize startTransfer;
 
     public void subtract(InformationSize informationSizeToSubtract) {
-        Float actualTransferInBytes = transfer.parseTo(InformationUnit.BYTE).size();
-        Float actualFileSizeInBytes = informationSizeToSubtract.parseTo(InformationUnit.BYTE).size();
+        Long actualTransferInBytes = transfer.parseTo(InformationUnit.BYTE).size();
+        Long actualFileSizeInBytes = informationSizeToSubtract.parseTo(InformationUnit.BYTE).size();
 
-        Float diff = actualTransferInBytes - actualFileSizeInBytes;
+        Long diff = actualTransferInBytes - actualFileSizeInBytes;
         InformationSize diffInformationSize = new InformationSize(InformationUnit.BYTE, diff);
 
         this.transfer = diffInformationSize.parseTo(transfer.unit());
